[FS] Regenerated demo for fs__qspi__app_programmer__firmware__pic32ck_gc01_cult.X
MH3-113065

+++ apps/fs/qspi/app_programmer/firmware/src/config/pic32ck_gc01_cult/driver/sdspi/src/drv_sdspi.c
@@ -2528,7 +2528,7 @@ static void DRV_SDSPI_SetupXfer(
         {
             return;
         }
-        if (((blockStart + nBlocks) > dObj->mediaGeometryTable[SYS_MEDIA_GEOMETRY_TABLE_READ_ENTRY].numBlocks))
+        if ((((uint64_t)blockStart + nBlocks) > dObj->mediaGeometryTable[SYS_MEDIA_GEOMETRY_TABLE_READ_ENTRY].numBlocks))
         {
             return;
         }
@@ -2539,7 +2539,7 @@ static void DRV_SDSPI_SetupXfer(
         {
             return;
         }
-        if (((blockStart + nBlocks) > dObj->mediaGeometryTable[SYS_MEDIA_GEOMETRY_TABLE_WRITE_ENTRY].numBlocks))
+        if ((((uint64_t)blockStart + nBlocks) > dObj->mediaGeometryTable[SYS_MEDIA_GEOMETRY_TABLE_WRITE_ENTRY].numBlocks))
         {
             return;
         }
